First Prototype ready, barring bugs
I switched all the doubles to floats, and renamed the
record_measurements method to start_distance_process. I also changed it
in the xml, so pressing the button should still work as it did before.

I also changed some of the measurement logic, but not really. The only
thing I did was instead of having it wait on some semaphore structure
that I may not actually understand, I implemented the wait using a
CountDownLatch. Think of this as a gate that is closed until the
countDown() method has been invoked enough times (coder-specified).
Processes/threads await() on the gate, and when the countdown has been
reached, all processes/threads awaken.