App "works"
I couldn't get the emulator or the real phone to register a listener of
type LINEAR_ACCELERATION. This is a newer sensor type, so it is
possible that it just doesn't exist aas far as my phone and emulator
are concerned.

Instead, we will have to account for gravity using either a high-pass
filter as suggested here:
http://developer.android.com/reference/android/hardware/SensorEvent.html

Or, we can use Sensor.getOrientation() and construct our own gravity
vector based on the orientation of the phone. If we do that, we will
need to also record the phone's orientation at measure-time, and
construct gravity between measure-time and distance calculation.

Other than that, we have a working app that gives "correct" data,
assuming the phone is falling toward the earth.